+++ example/android/build.gradle
@@ -2,6 +2,7 @@
 
 buildscript {
     ext {
+        kotlinVersion = "1.6.0"
         buildToolsVersion = "31.0.0"
         minSdkVersion = 21
         compileSdkVersion = 31
