+++ example/package.json
@@ -11,8 +11,15 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
     "react": "18.1.0",
     "react-native": "0.70.6",
+    "react-native-gesture-handler": "^2.9.0",
+    "react-native-reanimated": "^3.1.0",
+    "react-native-safe-area-context": "^4.5.2",
+    "react-native-screens": "^3.20.0",
     "sift-react-native": "^0.1.8"
   },
   "devDependencies": {
