+++ package.json
@@ -23,7 +23,7 @@
     "test": "jest",
     "typescript": "tsc --noEmit",
     "lint": "eslint \"**/*.{js,ts,tsx}\"",
-    "prepare": "bob build",
+    "prepare": "bob build && husky install",
     "release": "release-it",
     "example": "yarn --cwd example",
     "pods": "cd example && pod-install --quiet",
@@ -50,10 +50,10 @@
     "@types/react": "^16.9.19",
     "@types/react-native": "^0.70.6",
     "commitlint": "^8.3.5",
-    "eslint": "^7.2.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.3",
-    "husky": "^4.2.5",
+    "husky": "^8.0.0",
     "jest": "^26.0.1",
     "pod-install": "^0.1.0",
     "prettier": "^2.0.5",
@@ -72,11 +72,6 @@
       "<rootDir>/example/node_modules",
       "<rootDir>/lib/"
     ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "yarn lint && yarn typescript"
-    }
   },
   "commitlint": {
     "extends": [
@@ -138,5 +133,9 @@
       "module",
       "typescript"
     ]
+  },
+  "dependencies": {
+    "@react-navigation/native-stack": "^6.9.12",
+    "sift-react-native": "^0.1.9"
   }
 }
